% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_hypothesis_test.R
\name{argmin.HT.bootstrap}
\alias{argmin.HT.bootstrap}
\title{Perform argmin hypothesis test.}
\usage{
argmin.HT.bootstrap(data, r, sample.mean = NULL, alpha = 0.05, B = 200)
}
\arguments{
\item{data}{A n by p data matrix; each of its row is a p-dimensional sample.}

\item{r}{The dimension of interest for hypothesis test.}

\item{sample.mean}{The sample mean of the n samples in data; defaults to NULL. It can be calculated via colMeans(data).
If your experiment involves hypothesis testing over more than one dimension, compute colMeans(data) and pass it to sample.mean to speed up computation.}

\item{alpha}{The significance level of the hypothesis test; defaults to 0.05.}

\item{B}{The number of bootstrap samples; defaults to 200.}
}
\value{
A list containing:\tabular{ll}{
   \code{p.val} \tab p value of the test \cr
   \tab \cr
   \code{test.stat} \tab max statistic \cr
   \tab \cr
   \code{quantile} \tab (1-alpha) quantile of the bootstrap samples \cr
   \tab \cr
   \code{ans} \tab 'Reject' or 'Accept' \cr
}
}
\description{
Test if a dimension may be argmin, adapting the one-step bootstrap method in \insertCite{cck.many.moments}{argminCS}.
}
\references{
{
 \insertRef{cck.many.moments}{argminCS}

 \insertRef{lei.cvc}{argminCS}
}
}
