% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_hypothesis_test.R
\name{argmin.HT.fold}
\alias{argmin.HT.fold}
\title{Perform argmin hypothesis test.}
\usage{
argmin.HT.fold(data, r, lambda, alpha = 0.05, n.fold = 2)
}
\arguments{
\item{data}{A n by p data matrix; each of its row is a p-dimensional sample.}

\item{r}{The dimension of interest for hypothesis test.}

\item{lambda}{The real-valued tuning parameter for exponential weightings (the calculation of softmin).}

\item{alpha}{The significance level of the hypothesis test; defaults to 0.05.}

\item{n.fold}{The number of folds.}
}
\value{
A list containing:\tabular{ll}{
   \code{test.stat.scale} \tab The scaled test statistic \cr
   \tab \cr
.   \code{critical.value} \tab The critical value for the hypothesis test. Being greater than it leads to a rejction. \cr
   \tab \cr
   \code{std} \tab The standard deviation estimate. \cr
   \tab \cr
   \code{ans} \tab 'Reject' or 'Accept' \cr
}
}
\description{
Test if a dimension may be argmin by splitting into folds.
}
\details{
This method does not support any data-driven way to tune lambda for now. Try to use argmin.HT.LOO instead.
}
