% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_hypothesis_test.R
\name{argmin.HT.GU}
\alias{argmin.HT.GU}
\title{Perform argmin hypothesis test.}
\usage{
argmin.HT.GU(
  data,
  r,
  omega = NULL,
  estimated.minimum.mean = NULL,
  mean.r = NULL,
  alpha = 0.05,
  ...
)
}
\arguments{
\item{data}{A n by p data matrix; each of its row is a p-dimensional sample.}

\item{r}{The dimension of interest for hypothesis test.}

\item{omega}{The tuning parameter in the statistic; defaults to NULL. Recommend to compute it using \link{omega.bootstrap}.}

\item{estimated.minimum.mean}{Upon a splitting over data, the mean of r'-th dimension over the testing set, where r' is the argmin over the training set; defaults to NULL.
Highly recommend to provide a value to improve the computation speed.}

\item{mean.r}{Upon a splitting over data, the mean of the r-th dimension over the testing set; defaults to NULL.
Highly recommend to provide a value to improve the computation speed.}

\item{alpha}{The significance level of the hypothesis test; defaults to 0.05.}

\item{...}{Additional arguments to \link{omega.bootstrap}.}
}
\value{
A list containing:\tabular{ll}{
   \code{test.statistic} \tab The off-line test statistic in \insertCite{dey.2024}{argminCS}. \cr
   \tab \cr
.   \code{critical.value} \tab The critical value for the hypothesis test. Being greater than it leads to a rejection. \cr
   \tab \cr
   \code{ans} \tab 'Reject' or 'Accept' \cr
}
}
\description{
Test if a dimension may be argmin, employing the universal inference framework by \insertCite{dey.2024}{argminCS}.
}
\references{
{
  \insertRef{dey.2024}{argminCS}
}
}
