% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_hypothesis_test.R
\name{argmin.HT.MT}
\alias{argmin.HT.MT}
\title{Perform argmin hypothesis test.}
\usage{
argmin.HT.MT(data, r, test = "z", r.min = NULL, r.min.sec = NULL, alpha = 0.05)
}
\arguments{
\item{data}{A n by p data matrix; each of its row is a p-dimensional sample.}

\item{r}{The dimension of interest for hypothesis test.}

\item{test}{The test to perform: 't' or 'z' test; defaults to 'z'.
If data are believed to be normally distributed, use 't'; otherwise 'z'.}

\item{r.min}{The sample argmin of the sample mean of data; defaults to NULL.}

\item{r.min.sec}{The index of the second smallest dimension of the sample mean of the data; defaults to NULL.}

\item{alpha}{The significance level of the hypothesis test; defaults to 0.05.}
}
\value{
A list containing:\tabular{ll}{
   \code{p.val} \tab p value without Bonferroni's correction. \cr
   \tab \cr
.   \code{critical.value} \tab The critical value for the hypothesis test. Being greater than it leads to a rejction. \cr
   \tab \cr
   \code{ans} \tab 'Reject' or 'Accept' \cr
}
}
\description{
Test if a dimension may be argmin, using multiple testing with Bonferroni's correction.
}
