% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_hypothesis_test.R
\name{argmin.HT.nonsplit}
\alias{argmin.HT.nonsplit}
\title{Perform argmin hypothesis test.}
\usage{
argmin.HT.nonsplit(data, r, lambda, sample.mean = NULL, alpha = 0.05)
}
\arguments{
\item{data}{A n by p data matrix; each of its row is a p-dimensional sample.}

\item{r}{The dimension of interest for hypothesis test.}

\item{lambda}{The real-valued tuning parameter for exponential weightings (the calculation of softmin).}

\item{sample.mean}{The sample mean of the n samples in data; defaults to NULL. It can be calculated via colMeans(data).
If your experiment involves hypothesis testing over more than one dimension, pass sample.mean=colMeans(data) to speed up computation.}

\item{alpha}{The significance level of the hypothesis test; defaults to 0.05.}
}
\value{
A list containing:\tabular{ll}{
   \code{test.stat.scale} \tab The scaled test statistic \cr
   \tab \cr
.   \code{critical.value} \tab The critical value for the hypothesis test. Being greater than it leads to a rejction. \cr
   \tab \cr
   \code{std} \tab The standard deviation estimate. \cr
   \tab \cr
   \code{ans} \tab 'Reject' or 'Accept' \cr
}
}
\description{
Test if a dimension may be argmin without any splitting.
}
\details{
This method is not recommended, given its poor performance when p is small.
}
