% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_hypothesis_test.R
\name{argmin.HT.SN}
\alias{argmin.HT.SN}
\title{Perform argmin hypothesis test.}
\usage{
argmin.HT.SN(data, r, sample.mean = NULL, alpha = 0.05)
}
\arguments{
\item{data}{A n by p data matrix; each of its row is a p-dimensional sample.}

\item{r}{The dimension of interest for hypothesis test.}

\item{sample.mean}{The sample mean of the n samples in data; defaults to NULL. It can be calculated via colMeans(data).
If your experiment involves hypothesis testing over more than one dimension, compute colMeans(data) and pass it to sample.mean to speed up computation.}

\item{alpha}{The significance level of the hypothesis test; defaults to 0.05.}
}
\value{
A list containing:\tabular{ll}{
   \code{test.stat} \tab The test statistic \cr
   \tab \cr
.   \code{critical.value} \tab The critical value for the hypothesis test. Being greater than it leads to a rejection. \cr
   \tab \cr
   \code{ans} \tab 'Reject' or 'Accept' \cr
}
}
\description{
Test if a dimension may be argmin, adapting the one-step self-normalization method in \insertCite{cck.many.moments}{argminCS}.
}
\references{
{
 \insertRef{cck.many.moments}{argminCS}
}
}
