% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find.sub.argmin}
\alias{find.sub.argmin}
\title{Get the index of the smallest dimension apart from an index}
\usage{
find.sub.argmin(nums, idx, seed = NULL)
}
\arguments{
\item{nums}{A vector of numbers}

\item{idx}{An index to be excluded}

\item{seed}{(Optional) If provided, used to seed the random sampling (for reproducibility).}
}
\value{
The index of the second smallest dimension (as an integer).
}
\description{
Get the index of the smallest dimension apart from an index
}
\examples{
nums <- c(1,3,2)
find.sub.argmin(nums,1)
## return 3

nums <- c(1,1,2)
find.sub.argmin(nums,1)
## return 2
}
