% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_confidence_set.R
\name{get.sample.mean.r}
\alias{get.sample.mean.r}
\title{Compute sample mean differences for hypothesis testing}
\usage{
get.sample.mean.r(sample.mean, r)
}
\arguments{
\item{sample.mean}{A vector of length \code{p} containing the sample means of each dimension.}

\item{r}{An integer between 1 and \code{p}, indicating the reference dimension.}
}
\value{
A vector of length \code{p - 1} giving the differences: sample.mean[r] - sample.mean[-r].
}
\description{
Computes the vector of differences between the sample mean at the reference index
and the remaining dimensions.
}
\examples{
sample.mean <- 1:5
get.sample.mean.r(sample.mean, r = 3)

}
\keyword{internal}
