% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_test_statistic.R
\name{getMin.argmin.LOO}
\alias{getMin.argmin.LOO}
\title{Compute the minimum, using argmin.}
\usage{
getMin.argmin.LOO(
  i,
  r,
  data,
  lambda,
  sample.mean = NULL,
  ties.method = "average",
  seed = NULL
)
}
\arguments{
\item{i}{An index of the sample to be left out.}

\item{r}{The dimenion of interest for hypothesis test; excluded from the calculation of softmin.}

\item{data}{A n by p data matrix; each of its row is a p-dimensional sample.}

\item{sample.mean}{The sample mean of the n samples in data; defaults to be NULL. It can be calculated via colMeans(data).}

\item{ties.method}{A string indicating the method to tackle with ties: 'average' (or simply 'a'), 'random' (or simple 'r); defaults to average.}

\item{seed}{An integer seed in case that 'random' is chosen to tackle with ties; defaults to NULL.}
}
\value{
The minimum calculated from argmin under the leave-one-out scheme with the r-th dimension excluded (The quantity Q in Zhang et al).
}
\description{
Compute the minimum, using (hard)argmin, under the leave-one-out (LOO) scheme;
more specifically, the quantity Q in Zhang et al.
}
\examples{
n <- 100
mu <- (1:10)/10
cov <- diag(length(mu))
set.seed(31)
data <- MASS::mvrnorm(n, mu, cov)
## let the function calculate the sample.mean
getMin.argmin.LOO(7, 1, data, sqrt(n))

## calculate the sample.mean outside the function. This may foster LOO implementation computation-wise
getMin.argmin.LOO(7, 1, data, sqrt(n), sample.mean=colMeans(data))
}
