% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_driven_lambda.R
\name{lambda.adaptive.fold}
\alias{lambda.adaptive.fold}
\title{Generate a data-driven \eqn{\lambda} for fixed fold algorithm.}
\usage{
lambda.adaptive.fold(data, r, flds, const = 2.5)
}
\arguments{
\item{data}{A n by p data matrix; each of its row is a p-dimensional sample.}

\item{r}{The dimension of interest for hypothesis test.}

\item{flds}{A list of row position integers corresponding to folds.}

\item{const}{A scaling constant for the data driven \eqn{\lambda}; defaults to 2.5. As its value gets larger, the first order stability tends to increase while power might decrease.}
}
\value{
A data-driven \eqn{\lambda} for fixed fold algorithm.
}
\description{
Generate a data-driven \eqn{\lambda} for fixed fold algorithm motivated by the derivation of the first order stability.
}
\examples{
n <- 300
mu <- (1:10)/10
cov <- diag(length(mu))
set.seed(31)
data <- MASS::mvrnorm(n, mu, cov)
flds <- caret::createFolds(1:n, k=2, list=TRUE, returnTrain=FALSE)

## compute a data-driven lambda, and check its feasibility
lambda.adaptive.fold(data, 1, flds)

## want to ensure a greater stability
lambda.adaptive.fold(data, 1, flds, const=3)
}
