% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_driven_lambda.R
\name{lambda.adaptive}
\alias{lambda.adaptive}
\title{Generate a data-driven \eqn{\lambda}.}
\usage{
lambda.adaptive(data, r, sample.mean = NULL, const = 2.5)
}
\arguments{
\item{data}{A n by p data matrix; each of its row is a p-dimensional sample.}

\item{r}{The dimension of interest for hypothesis test.}

\item{sample.mean}{The sample mean of the n samples in data; defaults to NULL. It can be calculated via colMeans(data).}

\item{const}{A scaling constant for the data driven \eqn{\lambda}; defaults to 2.5. As its value gets larger, the first order stability tends to increase while power might decrease.}
}
\value{
A data-driven \eqn{\lambda}.
}
\description{
Generate a data-driven \eqn{\lambda} motivated by the derivation of the first order stability.
For its precise definition, we refer to the paper Zhang et al 2024.
}
\examples{
n <- 300
mu <- (1:10)/10
cov <- diag(length(mu))
set.seed(31)
data <- MASS::mvrnorm(n, mu, cov)
sample.mean <- colMeans(data)

## compute a data-driven lambda, and check its feasibility
lambda.adaptive(data, 1, sample.mean=sample.mean)

## want to ensure a greater stability
lambda.adaptive(data, 1, sample.mean=sample.mean, const=3)
}
