% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_hypothesis_test.R
\name{omega.bootstrap}
\alias{omega.bootstrap}
\title{tune the omega used for the universal inference method by \insertCite{dey.2024}{argminCS}.}
\usage{
omega.bootstrap(data, alpha = 0.05, B = 100, omegas = 1:100)
}
\arguments{
\item{data}{A n by p data matrix; each of its row is a p-dimensional sample.}

\item{alpha}{The significance level of the hypothesis test; defaults to 0.05.}

\item{B}{The number of bootstrap iterations; defaults to 100.}

\item{omegas}{A vector including the omegas to perform bootstrapping over; defaults to 1:100.}
}
\value{
An omega value for hypothesis testing, using the universal inference algorithm.
}
\description{
tune the omega used for the universal inference method by \insertCite{dey.2024}{argminCS} by bootstrapping.
}
\examples{
n <- 200
mu <- (1:20)/20
cov <- diag(length(mu))
set.seed(38)
data <- MASS::mvrnorm(n, mu, cov)
omega <- omega.bootstrap(data)
omega

}
\references{
{
  \insertRef{dey.2024}{argminCS}
}
}
