% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_hypothesis_test.R
\name{argmin.HT.fold}
\alias{argmin.HT.fold}
\title{Perform argmin hypothesis test.}
\usage{
argmin.HT.fold(
  data,
  r,
  alpha = 0.05,
  n.fold = 2,
  flds = NULL,
  sample.mean = NULL,
  min.algor = "softmin",
  lambda = NULL,
  const = 2.5,
  enlarge = TRUE,
  true.mean = NULL,
  output.weights = FALSE,
  ...
)
}
\arguments{
\item{data}{A n by p data matrix; each of its row is a p-dimensional sample.}

\item{r}{The dimension of interest for hypothesis test.}

\item{alpha}{The significance level of the hypothesis test; defaults to 0.05.}

\item{n.fold}{The number of folds.}

\item{flds}{A list of row position integers corresponding to folds; defaults to NULL.}

\item{sample.mean}{The sample mean of the n samples in data; defaults to NULL. It can be calculated via colMeans(data).
It is used to tune lambda in a data-driven way.}

\item{min.algor}{The algorithm to compute the minimum: 'softmin' or 'argmin'; defaults to 'softmin'}

\item{lambda}{The real-valued tuning parameter for exponential weightings (the calculation of softmin); defaults to NULL.
If lambda=NULL (recommended), the function would determine a lambda value in a data-driven way.}

\item{const}{A scaling constant for the data driven \eqn{\lambda}; defaults to 2.5. As its value gets larger, the first order stability tends to increase while power might decrease.}

\item{enlarge}{A boolean value indicating if the data-driven lambda should be determined via an iterative enlarging algorithm; defaults to TRUE.}

\item{true.mean}{The population mean landscape; defaults to NULL. If a vector were provided, the centered test statistic would be outputted.
It is only useful for a simulation purpose.}

\item{output.weights}{A boolean variable specifying whether the exponential weights should be outputted; defaults to NULL.}

\item{...}{Additional arguments to \link{lambda.adaptive.enlarge}, \link{is.lambda.feasible.fold}.}
}
\value{
A list containing:\tabular{ll}{
   \code{test.stat.scale} \tab The scaled test statistic \cr
   \tab \cr
.   \code{critical.value} \tab The critical value for the hypothesis test. Being greater than it leads to a rejection. \cr
   \tab \cr
   \code{std} \tab The standard deviation estimate. \cr
   \tab \cr
   \code{ans} \tab 'Reject' or 'Accept' \cr
   \tab \cr
   \code{lambda} \tab The lambda used in the hypothesis testing. \cr
   \tab \cr
   \code{test.stat.centered} \tab (Optional) The centered test statistic. Outputted only when true.mean is not NULL. \cr
   \tab \cr
   \code{exponential.weights} \tab (Optional) A (n.fold by p) matrix storing the exponential weightings in the test statistic. \cr
}
}
\description{
Test if a dimension may be argmin by splitting into folds.
}
\details{
This method does not support any data-driven way to tune lambda for now. Try to use argmin.HT.LOO instead.
}
