% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_hypothesis_test.R
\name{argmin.HT.LOO}
\alias{argmin.HT.LOO}
\title{Perform argmin hypothesis test.}
\usage{
argmin.HT.LOO(
  data,
  r,
  sample.mean = NULL,
  min.algor = getMin.softmin.LOO,
  lambda = NULL,
  const = 2.5,
  enlarge = TRUE,
  alpha = 0.05,
  ...
)
}
\arguments{
\item{data}{A n by p data matrix; each of its row is a p-dimensional sample.}

\item{r}{The dimension of interest for hypothesis test.}

\item{sample.mean}{The sample mean of the n samples in data; defaults to NULL. It can be calculated via colMeans(data).
If your experiment involves hypothesis testing over more than one dimension, compute colMeans(data) and pass it to sample.mean to speed up computation.}

\item{min.algor}{The algorithm to find the minimum excluding the r-th dimension; defaults to \link{getMin.softmin.LOO}. The other option is \link{getMin.argmin.LOO}.}

\item{lambda}{The real-valued tuning parameter for exponential weightings (the calculation of softmin); defaults to NULL.
If lambda=NULL (recommended), the function would determine a lambda value in a data-driven way.}

\item{const}{A scaling constant for the data driven \eqn{\lambda}; defaults to 2.5. As its value gets larger, the first order stability tends to increase while power might decrease.}

\item{enlarge}{A boolean value indicating if the data-driven lambda should be determined via an iterative enlarging algorithm; defaults to TRUE.}

\item{alpha}{The significance level of the hypothesis test; defaults to 0.05.}

\item{...}{Additional arguments to \link{lambda.adaptive.enlarge}, \link{is.lambda.feasible}.}
}
\value{
A list containing:\tabular{ll}{
   \code{test.stat.scale} \tab The scaled test statistic \cr
   \tab \cr
.   \code{critical.value} \tab The critical value for the hypothesis test. Being greater than it leads to a rejection. \cr
   \tab \cr
   \code{std} \tab The standard deviation estimate. \cr
   \tab \cr
   \code{ans} \tab 'Reject' or 'Accept' \cr
}
}
\description{
Test if a dimension may be argmin, using the LOO (leave-one-out) algorithm in Zhang et al 2024.
}
