% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_hypothesis_test.R
\name{get.quantile.gupta.selection}
\alias{get.quantile.gupta.selection}
\title{Generate the quantile used for the selection procedure in \insertCite{gupta.1965}{argminCS}.}
\usage{
get.quantile.gupta.selection(p, alpha = 0.05, B = 1e+05)
}
\arguments{
\item{p}{The number of dimensions in your data matrix}

\item{alpha}{The level of the upper quantile; defaults to 0.05 (95 percentile).}

\item{B}{The number of Monte Carlo repetitions; defaults to 100000.}
}
\value{
A list containing:\tabular{ll}{
   \code{critica.val} \tab The 1 - alpha \cr
}
}
\description{
Generate the quantile used for the selection procedure in \insertCite{gupta.1965}{argminCS} by Monte Carlo estimation.
}
\examples{
p <- 10
get.quantile.gupta.selection(p, B=10000)

get.quantile.gupta.selection(p)

}
\references{
{
 \insertRef{gupta.1965}{argminCS}

 \insertRef{futschik.1995}{argminCS}
}
}
