% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_test_statistic.R
\name{getMin.softmin.LOO}
\alias{getMin.softmin.LOO}
\title{Compute the minimum, using softmin.}
\usage{
getMin.softmin.LOO(
  i,
  r,
  data,
  lambda,
  sample.mean = NULL,
  print.weights = FALSE
)
}
\arguments{
\item{i}{An index of the sample to be left out.}

\item{r}{The dimenion of interest for hypothesis test; excluded from the calculation of softmin.}

\item{data}{A n by p data matrix; each of its row is a p-dimensional sample.}

\item{lambda}{The real-valued tuning parameter for exponential weightings (the calculation of softmin).}

\item{sample.mean}{The sample mean of the n samples in data; defaults to be NULL. It can be calculated via colMeans(data).}

\item{print.weights}{A boolean specifying if the exponential weightings are printed to the console; defaults to False.}
}
\value{
The minimum calculated from softmin under the leave-one-out scheme with the r-th dimension excluded (The quantity Q in Zhang et al).
}
\description{
Compute the minimum, using softmin, under the leave-one-out (LOO) scheme;
more specifically, the quantity Q in Zhang et al.
}
\examples{
n <- 100
mu <- (1:10)/10
cov <- diag(length(mu))
set.seed(31)
data <- MASS::mvrnorm(n, mu, cov)
## let the function calculate the sample.mean
getMin.softmin.LOO(7, 1, data, sqrt(n))

## calculate the sample.mean outside the function. This may foster LOO implementation computation-wise
getMin.softmin.LOO(7, 1, data, sqrt(n), sample.mean=colMeans(data))

getMin.softmin.LOO(7, 1, data, sqrt(n), sample.mean=colMeans(data), print.weights=TRUE)
}
